pyproject.toml
[tool.poetry]
name = "ai-newsletter-backend"
version = "0.1.0"
description = ""
authors = ["junlee <laonzenamoon@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
fastapi = "^0.115.6"
uvicorn = "^0.34.0"
langsmith = "^0.2.4"
langgraph = "^0.2.60"
langchain-openai = "^0.2.14"
python-dotenv = "^1.0.1"
aiomysql = "^0.2.0"
redis = "^5.2.1"
tsidpy = "^1.1.5"
aiohttp = "^3.11.11"
sqlalchemy = "^2.0.36"
greenlet = "^3.1.1"
pytest = "^8.3.4"
ipython = "^8.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
